src/requirements.txt
arrow==1.2.3
asgiref==3.7.2
Babel==2.12.1
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.3.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-browser-reload==1.11.0
django-environ==0.11.2
django-import-export==3.3.1
django-phonenumber-field==7.1.0
django-tailwind==3.6.0
et-xmlfile==1.1.0
idna==3.4
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
mdurl==0.1.2
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.20
Pillow==10.0.1
psycopg2==2.9.7
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.5.3
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.4
xlrd==2.0.1
xlwt==1.3.0
